Try fixing puppet_apply run

+++ puppet/scripts/install_puppet.sh
@@ -1,4 +1,13 @@
-wget https://apt.puppetlabs.com/puppetlabs-release-pc1-xenial.deb
-sudo dpkg -i puppetlabs-release-pc1-xenial.deb
-sudo apt update
-sudo apt install puppet-agent
+#!/usr/bin/env bash
+
+set -e
+
+REPO_DEB="puppetlabs-release-pc1-xenial.deb"
+if ! dpkg -s puppet-agent 2>1&>/dev/null; then
+    echo "Installing puppet"
+    wget https://apt.puppetlabs.com/${REPO_DEB}
+    sudo dpkg -i ${REPO_DEB}
+    rm ${REPO_DEB}
+    sudo apt update
+    sudo apt install puppet-agent
+fi
